+++ src/main/scala/com/axtstar/asta4e/converter/CC.scala
@@ -375,19 +375,5 @@ object CC {
     implicit val hnilFromMap: FromMap[HNil] = new FromMap[HNil] {
       def apply(m: Map[String, Any]): Option[HNil] = Some(HNil)
     }
-
-    implicit def hconsFromMap0[K <: Symbol, V, R <: HList, T <: HList](implicit
-                                                                       witness: Witness.Aux[K],
-                                                                       gen: LabelledGeneric.Aux[V, R],
-                                                                       fromMapH: FromMap[R],
-                                                                       fromMapT: FromMap[T]
-                                                                      ): FromMap[FieldType[K, V] :: T] = new FromMap[FieldType[K, V] :: T] {
-      def apply(m: Map[String, Any]): Option[FieldType[K, V] :: T] = for {
-        v <- m.get(witness.value.name)
-        r <- Typeable[Map[String, Any]].cast(v)
-        h <- fromMapH(r)
-        t <- fromMapT(m)
-      } yield field[K](gen.from(h)) :: t
-    }
   }
 }
