+++ src/test/scala/com/axtstar/asta4e/ExcelBasicTest.scala
@@ -4,9 +4,7 @@ import java.io.File
 import java.text.SimpleDateFormat
 import java.util.Date
 
-import com.axtstar.asta4e.core.{ExcelBasic, MapHelper}
-import com.axtstar.asta4e.core.MapHelper.MapHelper
-import com.axtstar.asta4e.test_class.VariousCell
+import com.axtstar.asta4e.core.ExcelBasic
 import org.junit.runner.RunWith
 import org.specs2.mutable.Specification
 import org.specs2.runner.JUnitRunner
