+++ src/test/scala/com/axtstar/asta4e/MapHelperTest.scala
@@ -4,8 +4,7 @@ import java.io.File
 import java.text.SimpleDateFormat
 import java.util.Date
 
-import com.axtstar.asta4e.core.MapHelper
-import com.axtstar.asta4e.core.MapHelper.MapHelper
+import com.axtstar.asta4e.converter._
 import com.axtstar.asta4e.test_class.{Data_HOLIZONTAL, VariousCell}
 import org.junit.runner.RunWith
 import org.specs2.mutable.Specification
